# backend/docs/swagger.yaml
basePath: /
definitions:
  res.ErrorConstant:
    properties:
      code:
        type: integer
      data: {}
      error_fields:
        items:
          $ref: '#/definitions/res.Message'
        type: array
      message:
        type: string
      status:
        type: boolean
    type: object
  res.Message:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  res.ResponsePageConstant:
    properties:
      code:
        type: integer
      data: {}
      error_fields:
        items:
          $ref: '#/definitions/res.Message'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/utils.PaginationInfoDTO'
      status:
        type: boolean
    type: object
...
  schemas.CreateProduct:
    properties:
      brand:
        maxLength: 100
        type: string
      category_id:
        type: string
      description:
        type: string
      discount_price:
        minimum: 0
        type: number
      gender:
        enum:
        - male
        - female
        - unisex
        type: string
      is_active:
        type: boolean
      material:
        maxLength: 150
        type: string
      price:
        type: number
      product_name:
        maxLength: 255
        minLength: 3
        type: string
      product_slug:
        maxLength: 255
        type: string
      seo_description:
        type: string
      seo_title:
        maxLength: 255
        type: string
      specifications:
        items:
          $ref: '#/definitions/schemas.CreateProductSpecification'
        type: array
      status:
        enum:
        - active
        - draft
        - archived
        type: string
      stock:
        minimum: 0
        type: integer
      variants:
        items:
          $ref: '#/definitions/schemas.CreateProductVariant'
        type: array
      weight:
        minimum: 0
        type: number
    required:
    - category_id
    - price
    - product_name
    type: object
  schemas.CreateProductSpecification:
    properties:
      spec_name:
        type: string
      spec_value:
        type: string
    required:
    - spec_name
    - spec_value
    type: object
  schemas.CreateProductVariant:
    properties:
      barcode:
        type: string
      color:
        type: string
      is_active:
        type: boolean
      price_adjustment:
        type: number
      size:
        type: string
      sku:
        type: string
      stock:
        type: integer
      weight:
        type: number
    type: object
...
  schemas.UpdateProduct:
    properties:
      brand:
        maxLength: 100
        type: string
      category_id:
        type: string
      description:
        type: string
      discount_price:
        minimum: 0
        type: number
      gender:
        enum:
        - male
        - female
        - unisex
        type: string
      is_active:
        type: boolean
      material:
        maxLength: 150
        type: string
      price:
        type: number
      product_name:
        maxLength: 255
        minLength: 3
        type: string
      product_slug:
        maxLength: 255
        type: string
      seo_description:
        type: string
      seo_title:
        maxLength: 255
        type: string
      status:
        enum:
        - active
        - draft
        - archived
        type: string
      stock:
        minimum: 0
        type: integer
      weight:
        minimum: 0
        type: number
    type: object
  schemas.UpdateProductVariant:
    properties:
      barcode:
        type: string
      color:
        type: string
      is_active:
        type: boolean
      price_adjustment:
        type: number
      size:
        type: string
      sku:
        type: string
      stock:
        type: integer
      weight:
        type: number
    type: object
...
  utils.PaginationInfoDTO:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: E-Commerce NW SBO Backend API
  title: E-Commerce NW API
  version: "1.0"
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: Get paginated list of products with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        required: true
        type: integer
      - default: 12
        description: Page size
        in: query
        name: page_size
        required: true
        type: integer
      - description: Search by name or brand
        in: query
        name: search
        type: string
      - description: Filter by category UUID
        in: query
        name: category_id
        type: string
      - description: Filter by brand
        in: query
        name: brand
        type: string
      - description: Filter by gender (male|female|unisex)
        in: query
        name: gender
        type: string
      - description: Filter by status (active|draft|archived)
        in: query
        name: status
        type: string
      - description: Minimum price
        in: query
        name: min_price
        type: number
      - description: Maximum price
        in: query
        name: max_price
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ResponsePageConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: GET Products (paginated)
      tags:
      - Products
  /products/{slug}:
    get:
      consumes:
      - application/json
      description: Get product detail by slug (public)
      parameters:
      - description: Product Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: GET Product by Slug
      tags:
      - Products
  /products/admin:
    post:
      consumes:
      - application/json
      description: Create a new product with optional variants and specifications
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateProduct'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: POST Create Product (Admin)
      tags:
      - Products
  /products/admin/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a product
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: DELETE Product (Admin)
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Get product detail by ID
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: GET Product by ID (Admin)
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update an existing product
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: PUT Update Product (Admin)
      tags:
      - Products
  /products/admin/{id}/images:
    get:
      consumes:
      - application/json
      description: Get all images for a product
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: GET Product Images (Admin)
      tags:
      - Products
    post:
      consumes:
      - multipart/form-data
      description: Upload product image to Cloudinary and save URL to DB
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Image file (jpg, png, webp, gif, bmp)
        in: formData
        name: image
        required: true
        type: file
      - description: 'Set as thumbnail (default: false)'
        in: formData
        name: is_thumbnail
        type: boolean
      - description: 'Sort order (default: 0)'
        in: formData
        name: sort_order
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: POST Upload Product Image (Admin)
      tags:
      - Products
  /products/admin/{id}/images/{image_id}:
    delete:
      consumes:
      - application/json
      description: Delete a product image from DB and Cloudinary
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Image ID (UUID)
        in: path
        name: image_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: DELETE Product Image (Admin)
      tags:
      - Products
  /products/admin/{id}/images/{image_id}/thumbnail:
    put:
      consumes:
      - application/json
      description: Set a specific image as the product thumbnail
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Image ID (UUID)
        in: path
        name: image_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: PUT Set Thumbnail Image (Admin)
      tags:
      - Products
  /products/admin/{id}/specifications:
    post:
      consumes:
      - application/json
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Spec data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateProductSpecification'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: POST Create Specification (Admin)
      tags:
      - Products
  /products/admin/{id}/specifications/{spec_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Specification ID (UUID)
        in: path
        name: spec_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: DELETE Specification (Admin)
      tags:
      - Products
  /products/admin/{id}/variants:
    post:
      consumes:
      - application/json
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Variant data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateProductVariant'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: POST Create Variant (Admin)
      tags:
      - Products
  /products/admin/{id}/variants/{variant_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Variant ID (UUID)
        in: path
        name: variant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: DELETE Variant (Admin)
      tags:
      - Products
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Variant ID (UUID)
        in: path
        name: variant_id
        required: true
        type: string
      - description: Variant data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateProductVariant'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: PUT Update Variant (Admin)
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"